refactor(web): extract clipboard hook and item subcomponents from context-picker
Split context-picker.tsx (609 → 441 lines) by extracting:
- use-context-picker-clipboard.ts (142 lines): all clipboard state, refs, and
  handlers (checkClipboard, pasteInput, pasteFeedback) into a custom hook
- context-picker-items.tsx (85 lines): ContextFileItem and ContextLinkItem
  subcomponents for file thumbnail and link badge rendering

Also deduplicated the popover content JSX that was repeated in both the
empty-state and has-items branches.
